chore(#4): switch to TypeScript
