+++ .gitlab-ci.yml
@@ -1,30 +1,21 @@
 image: node:10
 
 stages:
-- test
 - build
-
-test_local:
-    stage: test
-    script:
-    - npm install
-    - npm test
-    cache:
-        key: "$CI_BUILD_STAGE"
-        paths:
-        - node_modules/
+- test
 
 build_lib:
     stage: build
     script:
     - npm install
+    - npm run lint
     - npm run build
     artifacts:
         paths:
         - dist
         - lib
     cache:
-        key: "$CI_BUILD_STAGE"
+        key: "$CI_JOB_STAGE"
         paths:
         - node_modules/
 
@@ -35,8 +26,28 @@ build_doc:
     - npm run docs:build
     artifacts:
         paths:
-        - esdoc
+        - typedoc
     cache:
-        key: "$CI_BUILD_STAGE"
+        key: "$CI_JOB_STAGE"
+        paths:
+        - node_modules/
+
+test_lib:
+    stage: test
+    script:
+    - npm install
+    - npm run test:lib
+    cache:
+        key: "$CI_JOB_STAGE"
+        paths:
+        - node_modules/
+
+test_plugins:
+    stage: test
+    script:
+    - npm install
+    - npm run test:plugins
+    cache:
+        key: "$CI_JOB_STAGE"
         paths:
         - node_modules/
