--- .jshintignore
@@ -1,4 +1,0 @@
-esdoc
-dist
-lib
-node_modules
