--- esdoc.json
@@ -1,12 +1,0 @@
-{
-  "source": "./src",
-  "destination": "./esdoc",
-  "plugins": [
-    {
-      "name": "esdoc-standard-plugin",
-      "option": {
-        "title": "idomizer"
-      }
-    }
-  ]
-}
