+++ src/StringDictionary.ts
@@ -1,0 +1,3 @@
+export interface StringDictionary {
+    [type: string]: string
+}
