--- test/plugins/dummy.idomizer
@@ -1,1 +1,0 @@
-<h1 class="{{data.h1Class}}">Hello</h1>
